Move GeneralLogEntry to new package
