+++ src/main/java/teammates/ui/webapi/QueryLogsAction.java
@@ -8,9 +8,9 @@ import java.util.stream.Collectors;
 
 import org.apache.http.HttpStatus;
 
-import teammates.common.datatransfer.GeneralLogEntry;
 import teammates.common.datatransfer.QueryLogsParams;
 import teammates.common.datatransfer.QueryLogsResults;
+import teammates.common.datatransfer.logs.GeneralLogEntry;
 import teammates.common.datatransfer.logs.LogEvent;
 import teammates.common.datatransfer.logs.RequestLogUser;
 import teammates.common.datatransfer.logs.SourceLocation;
